requirements.txt
# TrendScope Analytics - NLP Trend Intelligence Pipeline
# Python Dependencies

# Core dependencies
requests==2.31.0
numpy==1.26.4
pandas==2.2.0

# NLP Processing
nltk==3.8.1

# Data Versioning
dvc==3.48.2
dvc-s3==3.2.0

# Workflow Orchestration
apache-airflow==2.8.1

# Additional utilities
python-dateutil==2.8.2
psutil==5.9.8
